nbiot collector v 0.8.5

+++ src/pages/advancedSettings.jsx
@@ -2,10 +2,33 @@ import { useState } from 'react'
 import { Page } from '../components/page'
 import "../styles/pages/advancedSettings.css"
 import global from '../store/global'
+import {setAdvSettings} from "../functions/requests";
+import {observer} from "mobx-react-lite";
 
-export default function AdvancedSettings() {
+export const AdvancedSettings =  observer (() => {
   const [settings, setSettings] = useState(localStorage.getItem("advSettings"));
-  const deflt = global.advSettings
+  const deflt = `[AppSettings]
+HostBroker=93.84.87.22
+PortBroker=1883
+GuardInterval=3
+AllPackages=1
+RawDataSave=1
+PSflag=true
+MqttPassword=1\\x10\\x17\\x12\\\\b\\\\a\\x1fP
+MqttUserName=admin
+
+[RemoteServices]
+AddRawDataUrl=http://192.168.0.201:5101/api/v2.1/rawdata/addv2
+LastParamInfoUrl="http://192.168.0.201:5101/api/v2.1/RawData/latest-param-info?source="
+LoginUrl=http://192.168.0.201:9109/api/v2.1/user/login
+Password=gateway
+RefreshTokenUrl=http://192.168.0.201:9109/api/v2.1/user/refresh-token
+UserInfoUrl=http://192.168.0.201:9109/api/v2.1/User/user-info
+UserName=gateway-3-11
+
+[TransferSource]
+MAC04_09_19_86_11_50=5896
+`
 
   function handleChange(e){
     setSettings(e.target.value)
@@ -13,21 +36,25 @@ export default function AdvancedSettings() {
   }
 
   function sendSettings() {
-    console.log("send:\n" + settings)
+    setAdvSettings(global.way + "/Advanced settings", settings, global.token)
   }
 
   function toDefault() {
-    let conf = window.confirm("вы уверены что хотите сбросить?")
+    let conf = window.confirm("вставить шаблон?")
     if(conf){
-      setSettings(deflt)
-      localStorage.setItem("advSettings", deflt)
+      if(settings) {
+        setSettings(settings + `\n\n` + deflt)
+      } else{
+        setSettings(settings + deflt)
+        localStorage.setItem("advSettings", deflt)
+      }
     }
   }
 
   return (
-      <Page 
-        header='Advanced settings' 
-        subHeader='Расширенные настройки' 
+      <Page
+        header='Advanced settings'
+        subHeader='Расширенные настройки'
         header2='Редактирование settings.ini'
         elem = {
           <>
@@ -36,9 +63,9 @@ export default function AdvancedSettings() {
               <textarea value={settings} onChange={(e) => handleChange(e)}/>
             </section>
             <button onClick={() => sendSettings()}>Сохранить</button>
-            <button onClick={() => toDefault()}>Сбросить</button>
+            <button onClick={() => toDefault()}>+ Шаблон</button>
           </>
         }
         />
   )
-}
+})
